SEQNG-271 Move code to handle emails on the security bundle to the ext bundle #time 1h
