+++ project/OcsApp.scala
@@ -157,5 +157,3 @@ trait OcsApp { this: OcsBundle =>
 
 }
 
-
-
